nixos/etebase-server: refactor hardcoded host to available user option
This commit refactors the hardcoded uvicorn host from 0.0.0.0 to the already
present user option, which defaults to 0.0.0.0.

This exists as the status quo does not respect the user's declared host to run
uvicorn on, i.e., `services.etebase-server.settings.allowed_hosts.allowed_host1'.
This user option is the "main host" address that the app should be accessible
from. However, the fact that uvicorn is started on 0.0.0.0 means that the app
will still remain accessible from all the other host addresses that the main
host may have; this may lead to an unexpected security issue.

One such example is the intention to host the app only on the local machine,
i.e., to set the user option to `localhost', and not have it accessible via the
machine's IP address, e.g., 192.168.1.10:8001. However, the app would still be
accessible at 192.168.1.10:8001, which goes against what the user intended in
the first place.