lib.linkedLists: init
Best way to return a large concatenated list from a `foldl'`, and
probably more use cases.

It's a means of last resort, but it's good to have somewhat of a soft
landing at least...
Engineering around the call stack was fun for me, so I figured I'd
just do it.