uboot*: add unique description to each board
If you search for "uboot" on https://search.nixos.org, the description is "Boot loader for embedded systems" for all boards, which is not very helpful.
it can be a bit difficult to tell from the Nixpkgs attribute name, which specific board a U-Boot variant is for.
I painstakingly collected the correct (AFAICT) board name (including upper/lower case, hyphens, spaces, etc.) and manufacturer for every single of the 63 currently in Nixpkgs supported boards, to make the description much more useful.
I chose to follow the pattern "U-Boot (boot loader) for the <Board> from <Manufacturer>" for most boards.
Exeptions are:
- `ubootTools` since these are generic tools not related to any specific hardware
- `ubootQemu*` where the definite article is not appropriate. And since it's a piece of software, there is also no separate "manufacturer"
- Boards like the Raspberry Pi and Orange Pi, since the board name already includes the manufacturer name and mentioning it again seems a bit weird
- Some boards like the Nvidia Jetson TK1, where it seems common to write the manufacturer name in front of the board name
- Boards by Pine64, because Pine64 uses something like "<board name> by Pine64" on their website