ghc_filesystem: disable new Clang 21 warning
This warning was added in Clang 21, causing this package to break on
Darwin as well as any other LLVM stdenv. Upstream hasn't changed and
hasn't released a version in two and a half years, so I'm just disabling
the warning.

(I did check whether updating to the latest `master` fixes the issue,
and it does not).