emacsPackages: don't set preBuild/patches to null
Depending on the version of the package from MELPA, nixpkgs can set
`preBuild` or `patches` to `null`. `mkDerivation` filters out `null`, so
this has the desirable effect of not unnecessarily changing the actual
derivation. But it is inconvenient when using `overrideAttrs`: it breaks
the common pattern of `(previousAttrs.preBuild or "") + ''...''` (also
used elsewhere in the files this commit changes), which will fail to
coerce `null` to a string or list to append to.

Fix it by defaulting `preBuild` to `""` and `patches` to `[ ]`.

We cannot set the attribute conditionally, because the attributes names
returned by `overrideAttrs` cannot depend on `finalAttrs` (only the
values can).