emacsPackages.bpr: fix future preBuild
We conditionally set `preBuild` depending on version. For a new enough
version of the `bpr` package, we would set `preBuild` to the
`previousAttrs` attrset.

This looks like a copy/paste error, so fix it.

This code is currently not reachable: it would need a newer version of
`bpr`, which has had no changes upstream for several years.