nixos/test-driver: refactor start-script creation
This change moves the instantiation of `NixStartScript` out into the
entry-point. That way, we can get rid of the iterator in `Drver.__init__`.

The main motivation is that this will be needed in the next commit to easily
pass VM names that are determined in NixStartScript to the function that
generates vhost_vsock sockets.