emacsPackages: don't set attributes to null
If the derivation they modify does not have packageRequires or
nativeBuildInputs attributes and the helper does not need to change
them, some helpers for applying selective overrides set these attributes
to null. `mkDerivation` filters out null (so this has the desirable
effect of not changing the actual derivation unnecessarily), but it is
inconvenient when using `overrideAttrs` on them: it breaks code like
`packageRequires = (previousAttrs.packageRequires or [ ]) ++ [
somePackage ]`.

Set these to the empty list instead.

We cannot set them conditionally, because the attribute names in
`overrideAttrs` cannot depend on `finalAttrs`, and the condition used by
these helpers commonly does depend on `finalAttrs`.`

It looks like this is effectively a noop, because the attributes
involved are always set (either by the helper or on the original
derivation).