nixos/sshguard: fix journalctl flag
The option description for `services.sshguard.services` claims that the
listed values are the services whose logs should be provided to
sshguard. However, the flag used with these values when provided to
journalctl means that they were considered the syslog identifiers.

Due to this issue, monitoring the sshd unit would not work by default.