mercurial: Use google-re2 instead
Facebook's fb-re2 hasn't been maintained for 6 years. Move Mercurial to
use google-re2, which is maintained, instead.

Mercurial has internal checks that make it work with either fb-re2 or
google-re2, so this should be fine.

As a bonus this also unblocks full Musl build of Nix, which depends on
Mercurial for tests.