or-tools: move pybind11-protobuf package into or-tools
or-tools is the only consumer of this library, and
1910563bb3c0a76aae035d578a8c094c5646baa7 is already one instance of the
toplevel build being broken.

Absorb pybind11-protobuf into or-tools, and manage the pinning there. It
will also make it much easier to do the upcoming or-tools bump, which
will also change these version bounds.

pybind11-protobuf is not actually packaged in the version that or-tools
asks for, but trying to use that version showed it wants abseil from
2023, which has been removed from nixpkgs.