qt6Packages.qxlsx: add fix for qt 6.10
`qxlsx` is failing to build with `nix-pkg-review`, with this error:

qxlsx> CMake Error at CMakeLists.txt:167 (target_link_libraries):
qxlsx>   Target "QXlsx" links to:
qxlsx>
qxlsx>     Qt6::GuiPrivate
qxlsx>
qxlsx>   but the target was not found.

The upstream `qxlsx` has been patched to fix this, but as a new upstream
version has not been released, this fix is not present in `nixpkgs` even
though `nixpkgs` has now bumped `qt6` to version 6.10.0 in PR 449572.

Thus, directly apply the patch made in this commit:
https://github.com/QtExcel/QXlsx/commit/90d762625750c6b2c73f6cd96b633e9158aed72e