python3Packages.kaleido: remove sbcl dependency (!!!)
This package was introduced in #339136, which attracted quite a bit of
conversation. I don't have the energy to go through the conversations
but it looks like some low-hanging fruit got overlooked. This package
depends on `sbcl` (!!) in order to get `fontconfig` files for `dejavu`
(!!!), which is ... already in the default `pkgs.fontconfig` config
files.

Also there was a bunch of crap in the `buildInputs` that wasn't needed
(like `bash` and packages which are interpolated in the `postInstall`).
I suspect that the `postInstall` is still pretty busted but this should
at least get `sbcl` out of my closure.