netlify-cli: 19.0.2 -> 23.9.2
Upgrades for netlify-cli were broken for a while because postinstall
script was changed to not build the packages into dist directory ahead
of time, but rather expect the user to do this. Add an workaround to not
run postinstall logic entirely.
Also disable edge functions in tests using an internal flag, because
they try to download Deno, some types definitions, and maybe other stuff
at runtime.