workflows/merge-group: prevent rebuild drift
If a package becomes more load-bearing between running a PR's Eval and
merging it, this can cause unexpected mass rebuilds. We call this a
"drift" in rebuilds and add a job to prevent this.

Small drift is OK, packages gain new dependencies all the time. For now,
we implement a simple check that tests whether a package caused fewer
than 1000 rebuilds when Eval last ran and then causes more than 1000
rebuilds once it hits the merge queue.

Certain improvements can and need to be made for this to work reliable
in various other scenarios, but these are considered out of scope for
now. Otherwise, I will never make it out of the deep rabbit hole I am
currently in.