nixos/immich: reindex VectorChord indexes on update
VectorChord requires its indexes to be reindexed when the extension is
updated. [1]
This commit adds functionality to save the current version of the
extension before performing an update, and then compare it with the
updated version to decide whether it should reindex Immich's indexes.
This complexity is needed to avoid reindexing every time PostgreSQL is
started, as it is an expensive operation that would slow down startup.

[1]: https://docs.immich.app/administration/postgres-standalone/#updating-vectorchord