python3Packages.tensorflow-probability: fix bazel build
Updated to build with newer versions of bazel.
tensorflow-probability upstream does not include a MODULE.bzl file so
added the `--enable_bzlmod=false` flag.
bazel also tries to fetch platforms and rules_cc from the internet
whilst building which is blocked by nix. These are now downloaded by
fetchzip and made available to the build with the --override_repository
flag