libbluray: 1.3.4 -> 1.4.0
This updates libbluray from version 1.3.4 to the latest v1.4.0. This
change is a bit more consequential than a typical version bump because
the underlying project switched from building with autotools to building
with meson in this release. Also, the underlying project is now
compatible with JDK 21, so the JDK version has been bumped.
Unfortunately, stable libbluray is not yet compatible with JDK 25
(changes that should hopefully make this work have been merged to their
mainline branch, but as of yet, no release has been made).

This change also introduces a new alternate version of libbluray into
nixpkgs, `libbluray-full`, which adds in all of the optional
dependencies, such as Java and libaacs. Nearly ten years ago, these
dependencies were removed from the default version of the package
because eelco didn't appreciate that a JDK was being installed on his
system, thus inflating the his closure. Although it is outside the scope
of this change, adding a "full" version of libbluray to nixpkgs should
allow downstream packages to consume it in the future and provide users
with full blu-ray support without having to build many of their packages
from scratch.