linux/build: re-expose `config` with its helper functions
In #448835 I renamed `config` to `configHelpers` to save a
`config_ = config;`, however `config` having the `is*` helpers is kinda
public API (as proven by NixOS modules failing to evaluate).

As a result, this commit adds the helpers back to `config` that is
exposed by the kernel derivation.

Apologies, I'm pretty sure I fixed that before locally and it just got
lost during a rebase :(