python3Packages.kaleido: remove sbcl dependency (!!!) (minimal)
This package was introduced in #339136, which attracted quite a bit of
conversation. I don't have the energy to go through the conversations
but it looks like some low-hanging fruit got overlooked. This package
depends on `sbcl` (!!) in order to get `fontconfig` files for `dejavu`
(!!!), which is ... already in the default `pkgs.fontconfig` config
files.

This is a stripped-down version of #454818 to (hopefully) expedite its
merge.