workflows/test: switch order of merge-group / pr jobs
The last commit made the PR workflow a dependency of the Merge Group
workflow, because the latter started to depend on artifacts of the
former to be available. This makes sense to do for the Test workflow,
because the PR workflow will always run before the Merge Queue in a real
setting.

This commit just adjusts the order in the test.yml file accordingly.
Separate commit to improve the diff.