quodlibet: don't use Xine unnecessarily
In 50dc0203707d (quodlibet: makeover, 2022-08-05) / #185236, the
quodlibet library was refactored.  That commit appears to have
accidentally changed the default arguments for building quodlibet to
mean the Xine backend would always be included, rather than only as an
alternative to the Gstreamer backend.  This adds unnecessary
dependencies for the non-Xine quodlibet packages, so default to only
including Xine if Gstreamer is absent.