sbcl: fix build on Darwin
Coreutils v9.8 has a regression on Darwin that causes
`__CF_USER_TEXT_ENCODING` env var to get added to `env`’s output. This
causes a test failure. So get the variant of `env` from
darwin.shell_cmds instead (and `uname`, to avoid having both pkgs in the
closure).

I submitted a bug report to Coreutils, but it is not yet in the tracker
or list archives for me to link from here.