workflows/merge-group: compare eval against previous queue item
This runs the compare step in the merge queue, to get a combined diff
against the previous queue item. This can be used as a base for multiple
different features in the future, for example preventing accidental mass
rebuilds in the merge queue, or diffing of packages' meta.

(cherry picked from commit 6ae5f33c8570507704f24920393e5f6ff3dbb441)