mesa: Only build OpenCL stuff on platforms where it works
Somewhere in OpenCL-related code (SPIRV-LLVM-Translator at minimum), things go wrong when running `mesa_clc` for OpenCL-using drivers on big-endian:

  Invalid extended instruction import ‘nepOs.LC’

Based on upstream issues and blog posts, the issue is likely in at least SPIRV-LLVM-Translator, so this one will have BE added to
its badPlatforms in the future.

When SPIRV-LLVM-Translator is unavailable, don't bother with any OpenCL stuff.