check-meta: allow `problems`, use in remediation
Waiting for a more comprehensive RFC 127 "Problems" implementation
(#177272), proactively packages to specify `meta.problems` and use
`problems` of kinds `unsupported` and `broken` to display context-aware
remediation messages.

This change is motivated by the merging of the CUDA13 PR, which included
denying in-tree support for using CUDA without enabling it Nixpkgs-wide.
Implementation-wise, unsupported packages were marked as `broken` (TBD:
make "unsupported"), and reasons are visible when `--trace-verbose`,
but obscured by the long and unhelpful NIXPKGS_ALLOW_BROKEN message.

Instead of reverting the CUDA13 changes, which are blocking for a number
of updates in the SciComp ecosystem, it seems better to allow
customization of remediation mesages.

In `cudaPackages` (and dependend `pythonXPackages`), we used to rely on
an ad hoc `brokenConditions` passthru attribute. Instead of moving that
to `meta`, rewrite them in the form compatible with the future RFC127
implementation.