firefox: relax disallowedRequisites to disallowedReferences to fix cuda build
When cudaSupport is enabled, cudaPackages.cuda_nvcc becomes a transitive
dependency of firefox.
It adds backendStdenv.cc to its propagatedBuildInputs which violates
firefox's disallowedRequisites policy.
This breaks firefox when cudaSupport is enabled.

I propose to relax the disallowedRequisites assertion and simply migrate
to disallowedReferences which ensures that stdenv.cc is not a direct
dependency of firefox.