pkgs.nixos: pass through `lib`
eval-config.nix re-imports lib/default.nix if it isn't passed an
instance which can bypass user defined extensions to lib. This is
confusing when overlaying nixpkgs with a modified lib and using
pkgs.nixos.

There are likely more instances of this sort of problem since things are
kind of happy to reimport lib, but I haven't spotted anything that seems
as serious. Most of those instances are jobset definitions and
maintainer scripts which should be fine to assume a vanilla lib.