emacsPackages: revert throw
We recently retired a lot of old aliases, but emacsPackages is still
exposed on search.nixos.org, so should not have been removed. I argued
previously, that this should not have been an alias in the first place:
If it's exposed on search.nixos.org, then it is expected to be used,
too.

We will also need that once we teach Eval to evaluate more packages than
hydra builds.