nixos/networkd: explicitly set IPv6AcceptRA=yes on networking.useDHCP=true
On a default NixOS installation, this setting is implied and there's no
real need to do it by yourself.

However, IPv6AcceptRA is disabled e.g. with IPv6 forwarding being
enabled. It took me a while to figure out why a machine only had an IPv4
address via DHCP, but didn't accept any RA.

It turned out that I had set

    systemd.network.config.networkConfig = {
      IPv6Forwarding = true;
    }

following upstream's recommendation[1] for setting it globally when IPv6
forwarding is desired. The result of silently turning off IPv6AcceptRA
was pretty unintuitive for me.

Since it's kinda the point of these definitions to be a fallback that
just enables automatic address allocation, I think it's reasonable to
set IPv6AcceptRA explicitly as well.

[1] https://github.com/yuwata/systemd/commit/865d314e4d294c6daa29d4ff482ca067df112316#diff-d81a981ca75657b8fc8f6ad6871cb9f8d7705944919b67dad745680ebeec1b4eR851