ci/github-script/merge: add trusted maintainers
This introduces the concept of "trusted maintainers", which are similar
to "scoped committers". They can merge changes from any contributor, as
long as these are scoped to the packages the TM maintains. Basic
criteria like by-name apply as well, of course.

It also allows the reverse flow, where a PR created by a trusted
maintainer can be merged by a regular maintainer of that package.

More elaborated checks for rebuild counts were intentionally left out,
because it is already the case that regular maintainers can merge
potentially breaking updates to load-bearing packages. To help *all*
maintainers deal with the fallout of such a merge, the merge-bot should
be extended to support merging reverts instead. This is out of scope for
here, though.