coreutils: fix regression in env on darwin
In v9.8, coreutils began linking against CoreFoundation. Doing so always
adds the `__CF_USER_TEXT_ENCODING` to the environment of the running
program. This means that `env -i env` is not empty, since the second
invocation adds the environment var. Backport the patch, but it is not
enough as this is a linking problem, and the upstream solution was to
avoid linking CoreFoundation in `env`, but when using
`--enable-single-binary`, it must be as everything is in one binary. As
such, disable and assert against enabling `singleBinary` on Darwin since
it results in incorrect behavior. FWIW, neither Homebrew nor MacPorts
enables single-binary.

See https://debbugs.gnu.org/cgi/bugreport.cgi?bug=79714