tests.fetchtorrent: verify and delete in postFetch
The Sintel torrent used for testing fetchtorrent is about 124MB, which
is downloaded and cached for every test every time the test inputs
change.  Avoid that by instead validating the file hashes in the
postFetch stage, then deleting the downloaded files, so the only thing
that needs to be stored is an empty directory.