gradle: pass through wrapper override arguments to unwrapped Gradle
Comments on #436358 and #402297 indicated a desire to have a more
controllable Gradle wrapper derivation. We were previously hiding the
Gradle being wrapped in the closure of wrapGradle, forcing us to expose
wrapGradle and use makeOverridable. Simply use callPackage to make it
easier for consumers to override things without us needing to expose gradleGen.