requireFile: allow substitutes
This was added in f194659ddbac889755737e7e2dc09b0bb821396a with the explanation "The requireFile
call was being substituted from the binary cache. We do not want this to happen"

As far as I can tell, no one has thought about this since. Preventing it here makes no sense:

- Anyone can "get around" this by doing `.overrideAttrs { allowSubstitutes = true; }`, and

- I *want* to cache requireFile derivations in my nixcache, and to do so I have to apply the above fix. I doubt I'm the only one.